Moved user language preferences into i18n.{h,cpp}
