+++ src/server/i18n.h
@@ -89,6 +89,16 @@ private: // data
   const Parameters  params;
 };
 
+struct LangPreference
+{
+  const std::string lang;
+  const float preference;
+};
+
+typedef std::vector<LangPreference> UserLangPreferences;
+
+UserLangPreferences parseUserLanguagePreferences(const std::string& s);
+
 } // namespace kiwix
 
 #endif // KIWIX_SERVER_I18N
